rename: change component folder name
