+++ app/(sub-page)/find-id/page.tsx
@@ -1,6 +1,6 @@
 import ContentContainer from '@/app/ui/view/atom/content-container/content-container';
 import TitleBox from '@/app/ui/view/molecule/title-box/title-box';
-import FindIdContainer from './[components]/find-id-container';
+import FindIdContainer from './components/find-id-container';
 import { Suspense } from 'react';
 import FindIdFormSkeleton from '@/app/ui/user/find-id-form/find-id-form.skeleton';
