+++ app/(sub-page)/find-password/page.tsx
@@ -1,6 +1,6 @@
 import ContentContainer from '@/app/ui/view/atom/content-container/content-container';
 import TitleBox from '@/app/ui/view/molecule/title-box/title-box';
-import FindPasswordContainer from './[components]/find-password-container';
+import FindPasswordContainer from './components/find-password-container';
 import { Suspense } from 'react';
 import FindIdFormSkeleton from '../../ui/user/find-id-form/find-id-form.skeleton';
